requirements.txt
selenium==4.16.0
behavex==3.2.0
allure-behave==2.8.18
allure-python-commons==2.8.18
jproperties
crypto
loguru
selenium-page-factory